Refactored assets client, api uses new assets client
