+++ .vscode/settings.json
@@ -26,6 +26,7 @@
     "Metaverse",
     "miniflare",
     "moonbase",
+    "moonsama",
     "nameprep",
     "nomicfoundation",
     "nomiclabs",
@@ -37,6 +38,7 @@
     "Polkadot",
     "QGIS",
     "roboto",
+    "Sama",
     "sepolia",
     "Smakhtin",
     "solhint",
