+++ packages/clients/assets-client/targets/directus/config.ts
@@ -1,0 +1,1 @@
+export const directusUrl = 'https://directus.sovereignnature.com';
