+++ packages/clients/assets-client/targets/directus/formatter.ts
@@ -1,0 +1,8 @@
+import { DeepAsset } from '../../types';
+import { DirectusAsset } from './types';
+
+export default function directusFormatter(assetData: DirectusAsset): DeepAsset {
+  const data = assetData.data;
+
+  return { ...data, tokenId: data.id };
+}
