+++ packages/clients/assets-client/targets/directus/types.ts
@@ -1,0 +1,13 @@
+export type DirectusAsset = {
+  data: {
+    id: string;
+    name: string;
+    description: string;
+    image: string;
+    collection: {
+      id: string;
+      name: string;
+      description: string;
+    };
+  };
+};
