+++ packages/clients/assets-client/targets/moonsama/client.ts
@@ -1,0 +1,20 @@
+import { GraphQLClient } from 'graphql-request';
+import { DeepAsset } from '../../types';
+import { moonSamaApiUrl } from './config';
+import { MoonSamaResponse } from './types';
+import { getNFT } from './queries';
+import moonSamaFormatter from './formatter';
+
+const moonSamaClient = new GraphQLClient(moonSamaApiUrl, { fetch });
+
+export async function getMoonSamaAsset(
+  contractAddress: string,
+  tokenId: number
+): Promise<DeepAsset> {
+  return moonSamaFormatter(
+    await moonSamaClient.request<MoonSamaResponse>(getNFT, {
+      contractAddress,
+      tokenId,
+    })
+  );
+}
