+++ packages/clients/assets-client/targets/moonsama/config.ts
@@ -1,0 +1,2 @@
+export const moonSamaApiUrl =
+  'https://squid.subsquid.io/moonsama-multichain/graphql';
