+++ packages/clients/assets-client/targets/moonsama/types.ts
@@ -1,0 +1,8 @@
+export type MoonSamaToken = {
+  id: string;
+  numericId: string;
+  address: string;
+  metadata: { name: string; description: string; image: string };
+};
+
+export type MoonSamaResponse = { data: { tokens: MoonSamaToken[] } };
