+++ packages/clients/assets-client/targets/polkadot/config.ts
@@ -1,0 +1,2 @@
+export const polkadotApiUrl = 'https://squid.subsquid.io/speck/graphql';
+export const kusamaApiUrl = 'https://squid.subsquid.io/stick/graphql';
