+++ packages/clients/assets-client/targets/polkadot/formatter.ts
@@ -1,0 +1,19 @@
+import { DeepAsset } from '../../types';
+import { PolkadotResponse } from './types';
+
+export default function polkadotFormatter(
+  assetData: PolkadotResponse
+): DeepAsset {
+  const nftEntity = assetData.nftEntity;
+  return {
+    id: nftEntity.id,
+    tokenId: nftEntity.sn,
+    name: nftEntity.meta.name,
+    description: nftEntity.meta.description,
+    image: nftEntity.meta.image,
+    collection: {
+      id: nftEntity.collection.id,
+      name: nftEntity.collection.name,
+    },
+  };
+}
