+++ packages/clients/assets-client/targets/polkadot/types.ts
@@ -1,0 +1,15 @@
+export type PolkadotResponse = {
+  nftEntity: {
+    id: string;
+    sn: string;
+    meta: {
+      name: string;
+      description: string;
+      image: string;
+    };
+    collection: {
+      id: string;
+      name: string;
+    };
+  };
+};
