+++ packages/clients/config.ts
@@ -1,4 +1,1 @@
-export const directusUrl = 'https://directus.sovereignnature.com';
-export const polkadotApiUrl = 'https://squid.subsquid.io/speck/graphql';
-export const kusamaApiUrl = 'https://squid.subsquid.io/stick/graphql';
 export const deepApiUrl = 'https://api.sovereignnature.com';
